requirements.txt
# requirements.txt
gitpython>=3.1.40
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.3.0
langchain==0.3.25
langchain-community>=0.3.24
langchain-core>=0.2.0
langchain-openai>=0.0.5
